fix: add toast when add product to cart and remove option to close sidebar by clicking outside it
